requirements.txt
psutil==5.8.0
PyYAML==5.4.1
requests==2.25.1
